requirements.txt
tda-api
discord.py
xmltodict
webdriver-manager